pyproject.toml
[project]
name = "chatgpt-proxy"
version = "0.1.0"
description = "Simple HTTP server that forwards text to ChatGPT and returns the reply."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.1",
    "mcp>=1.21.0",
    "openai>=1.51", # official OpenAI Python SDK (Responses API)
    "openai-agents>=0.0.9",
    "python-dotenv>=1.0",
    "requests>=2.32.5",
    "uvicorn>=0.30",
]